+++ tests/Core/Type/TypeConverterTest.php
@@ -73,7 +73,7 @@ class TypeConverterTest extends TestCase
     /**
      * @return TypeInterface[][]
      */
-    public function dataToFunctionType(): array
+    public function dataToExampleFnType(): array
     {
         return [
             [
@@ -119,7 +119,7 @@ class TypeConverterTest extends TestCase
             [new DoubleType(), new FloatType()],
             [new FalseType(), new BoolType()],
             [new MixedType(), null],
-            [new NullType(), null],
+            [new NullType(), new NullableType(new FqcnType('SomeClass'))],
             [new SelfType(), new SelfType()],
             [new StaticType(), null],
             [new ThisType(), null],
@@ -143,13 +143,14 @@ class TypeConverterTest extends TestCase
     }
 
     /**
-     * @dataProvider dataToFunctionType
+     * @dataProvider dataToExampleFnType
+     *
      * @param TypeInterface      $givenDocType
      * @param TypeInterface|null $expectedFnType
      */
-    public function testToFunctionType(TypeInterface $givenDocType, ?TypeInterface $expectedFnType): void
+    public function testToExampleFnType(TypeInterface $givenDocType, ?TypeInterface $expectedFnType): void
     {
-        $actualFnType = TypeConverter::toFunctionType($givenDocType);
+        $actualFnType = TypeConverter::toExampleFnType($givenDocType);
 
         self::assertEquals($expectedFnType, $actualFnType);
     }